requirements.txt
playwright-core==1.39.0
requests
